Move gunmod_remove(), gunmod_add(), gunmod_installation_odds(), toolmod_add()
