+++ src/player.cpp
@@ -91,10 +91,7 @@ static const efftype_id effect_blind( "blind" );
 static const efftype_id effect_downed( "downed" );
 static const efftype_id effect_stunned( "stunned" );
 
-static const itype_id itype_brass_catcher( "brass_catcher" );
-static const itype_id itype_large_repairkit( "large_repairkit" );
 static const itype_id itype_plut_cell( "plut_cell" );
-static const itype_id itype_small_repairkit( "small_repairkit" );
 
 static const trait_id trait_CF_HAIR( "CF_HAIR" );
 static const trait_id trait_DEBUG_HS( "DEBUG_HS" );
@@ -111,7 +108,6 @@ static const std::string flag_RESTRICT_HAND( "RESTRICT_HANDS" );
 
 static const skill_id skill_dodge( "dodge" );
 static const skill_id skill_gun( "gun" );
-static const skill_id skill_weapon( "weapon" );
 
 static const bionic_id bio_cqb( "bio_cqb" );
 
@@ -785,226 +781,6 @@ bool player::unload( item_location loc )
     add_msg( _( "You unload your %s." ), target->tname() );
     return true;
 
-}
-
-static bool has_mod( const item &gun, const item &mod )
-{
-    for( const item *toolmod : gun.gunmods() ) {
-        if( &mod == toolmod ) {
-            return true;
-        }
-    }
-    return false;
-}
-
-bool player::gunmod_remove( item &gun, item &mod )
-{
-    if( !has_mod( gun, mod ) ) {
-        debugmsg( "Cannot remove non-existent gunmod" );
-        return false;
-    }
-
-    item_location loc = item_location( *this, &mod );
-    if( mod.ammo_remaining() && !unload( loc ) ) {
-        return false;
-    }
-
-    gun.gun_set_mode( gun_mode_id( "DEFAULT" ) );
-    //TODO: add activity for removing gunmods
-
-    if( mod.typeId() == itype_brass_catcher ) {
-        gun.casings_handle( [&]( item & e ) {
-            return i_add_or_drop( e );
-        } );
-    }
-
-    const itype *modtype = mod.type;
-
-    i_add_or_drop( mod );
-    gun.remove_item( mod );
-
-    //If the removed gunmod added mod locations, check to see if any mods are in invalid locations
-    if( !modtype->gunmod->add_mod.empty() ) {
-        std::map<gunmod_location, int> mod_locations = gun.get_mod_locations();
-        for( const auto &slot : mod_locations ) {
-            int free_slots = gun.get_free_mod_locations( slot.first );
-
-            for( auto the_mod : gun.gunmods() ) {
-                if( the_mod->type->gunmod->location == slot.first && free_slots < 0 ) {
-                    gunmod_remove( gun, *the_mod );
-                    free_slots++;
-                } else if( mod_locations.find( the_mod->type->gunmod->location ) ==
-                           mod_locations.end() ) {
-                    gunmod_remove( gun, *the_mod );
-                }
-            }
-        }
-    }
-
-    //~ %1$s - gunmod, %2$s - gun.
-    add_msg_if_player( _( "You remove your %1$s from your %2$s." ), modtype->nname( 1 ),
-                       gun.tname() );
-
-    return true;
-}
-
-std::pair<int, int> player::gunmod_installation_odds( const item &gun, const item &mod ) const
-{
-    // Mods with INSTALL_DIFFICULT have a chance to fail, potentially damaging the gun
-    if( !mod.has_flag( "INSTALL_DIFFICULT" ) || has_trait( trait_DEBUG_HS ) ) {
-        return std::make_pair( 100, 0 );
-    }
-
-    int roll = 100; // chance of success (%)
-    int risk = 0;   // chance of failure (%)
-    int chances = 1; // start with 1 in 6 (~17% chance)
-
-    for( const auto &e : mod.type->min_skills ) {
-        // gain an additional chance for every level above the minimum requirement
-        skill_id sk = e.first == skill_weapon ? gun.gun_skill() : e.first;
-        chances += std::max( get_skill_level( sk ) - e.second, 0 );
-    }
-    // cap success from skill alone to 1 in 5 (~83% chance)
-    roll = std::min( static_cast<double>( chances ), 5.0 ) / 6.0 * 100;
-    // focus is either a penalty or bonus of at most +/-10%
-    roll += ( std::min( std::max( focus_pool, 140 ), 60 ) - 100 ) / 4;
-    // dexterity and intelligence give +/-2% for each point above or below 12
-    roll += ( get_dex() - 12 ) * 2;
-    roll += ( get_int() - 12 ) * 2;
-    // each level of damage to the base gun reduces success by 10%
-    roll -= std::max( gun.damage_level( 4 ), 0 ) * 10;
-    roll = std::min( std::max( roll, 0 ), 100 );
-
-    // risk of causing damage on failure increases with less durable guns
-    risk = ( 100 - roll ) * ( ( 10.0 - std::min( gun.type->gun->durability, 9 ) ) / 10.0 );
-
-    return std::make_pair( roll, risk );
-}
-
-void player::gunmod_add( item &gun, item &mod )
-{
-    if( !gun.is_gunmod_compatible( mod ).success() ) {
-        debugmsg( "Tried to add incompatible gunmod" );
-        return;
-    }
-
-    if( !has_item( gun ) && !has_item( mod ) ) {
-        debugmsg( "Tried gunmod installation but mod/gun not in player possession" );
-        return;
-    }
-
-    // first check at least the minimum requirements are met
-    if( !has_trait( trait_DEBUG_HS ) && !can_use( mod, gun ) ) {
-        return;
-    }
-
-    // any (optional) tool charges that are used during installation
-    auto odds = gunmod_installation_odds( gun, mod );
-    int roll = odds.first;
-    int risk = odds.second;
-
-    std::string tool;
-    int qty = 0;
-    bool requery = false;
-
-    item modded = gun;
-    modded.put_in( mod );
-    bool no_magazines = modded.common_ammo_default().is_null();
-
-    std::string query_msg = mod.is_irremovable()
-                            ? _( "<color_yellow>Permanently</color> install your %1$s in your %2$s?" )
-                            : _( "Attach your %1$s to your %2$s?" );
-    if( no_magazines ) {
-        query_msg += "\n";
-        query_msg += colorize(
-                         _( "Warning: after installing this mod, a magazine adapter mod will be required to load it!" ),
-                         c_red );
-    }
-
-    uilist prompt;
-    prompt.text = string_format( query_msg,
-                                 colorize( mod.tname(), mod.color_in_inventory() ),
-                                 colorize( gun.tname(), gun.color_in_inventory() ) );
-
-    std::vector<std::function<void()>> actions;
-
-    if( roll < 100 ) {
-        prompt.addentry( -1, true, 'w',
-                         string_format( _( "Try without tools (%i%%) risking damage (%i%%)" ), roll, risk ) );
-        actions.emplace_back( [&] {} );
-
-        prompt.addentry( -1, has_charges( itype_small_repairkit, 100 ), 'f',
-                         string_format( _( "Use 100 charges of firearm repair kit (%i%%)" ), std::min( roll * 2, 100 ) ) );
-
-        actions.emplace_back( [&] {
-            tool = "small_repairkit";
-            qty = 100;
-            roll *= 2; // firearm repair kit improves success...
-            risk /= 2; // ...and reduces the risk of damage upon failure
-        } );
-
-        prompt.addentry( -1, has_charges( itype_large_repairkit, 25 ), 'g',
-                         string_format( _( "Use 25 charges of gunsmith repair kit (%i%%)" ), std::min( roll * 3, 100 ) ) );
-
-        actions.emplace_back( [&] {
-            tool = "large_repairkit";
-            qty = 25;
-            roll *= 3; // gunsmith repair kit improves success markedly...
-            risk = 0;  // ...and entirely prevents damage upon failure
-        } );
-    } else {
-        prompt.addentry( -1, true, 'w', _( "Install without tools" ) );
-        actions.emplace_back( [&] {} );
-    }
-
-    prompt.addentry( -1, true, 'c', _( "Compare before/after installation" ) );
-    actions.emplace_back( [&] {
-        requery = true;
-        game_menus::inv::compare( gun, modded );
-    } );
-
-    do {
-        requery = false;
-        prompt.query();
-        if( prompt.ret < 0 ) {
-            add_msg_if_player( _( "Never mind." ) );
-            return; // player canceled installation
-        }
-        actions[ prompt.ret ]();
-    } while( requery );
-
-    const int moves = !has_trait( trait_DEBUG_HS ) ? mod.type->gunmod->install_time : 0;
-
-    assign_activity( activity_id( "ACT_GUNMOD_ADD" ), moves, -1, 0, tool );
-    activity.targets.push_back( item_location( *this, &gun ) );
-    activity.targets.push_back( item_location( *this, &mod ) );
-    activity.values.push_back( 0 ); // dummy value
-    activity.values.push_back( roll ); // chance of success (%)
-    activity.values.push_back( risk ); // chance of damage (%)
-    activity.values.push_back( qty ); // tool charges
-}
-
-void player::toolmod_add( item_location tool, item_location mod )
-{
-    if( !tool && !mod ) {
-        debugmsg( "Tried toolmod installation but mod/tool not available" );
-        return;
-    }
-    // first check at least the minimum requirements are met
-    if( !has_trait( trait_DEBUG_HS ) && !can_use( *mod, *tool ) ) {
-        return;
-    }
-
-    if( !query_yn( _( "Permanently install your %1$s in your %2$s?" ),
-                   colorize( mod->tname(), mod->color_in_inventory() ),
-                   colorize( tool->tname(), tool->color_in_inventory() ) ) ) {
-        add_msg_if_player( _( "Never mind." ) );
-        return; // player canceled installation
-    }
-
-    assign_activity( activity_id( "ACT_TOOLMOD_ADD" ), 1, -1 );
-    activity.targets.emplace_back( std::move( tool ) );
-    activity.targets.emplace_back( std::move( mod ) );
 }
 
 bool player::studied_all_recipes( const itype &book ) const
